pyproject.toml
[tool.poetry]
name = "event-planner-fastapi"
version = "0.1.0"
description = ""
authors = ["YoongSeong Lee <yslee@crowdworks.kr>"]
readme = "README.md"
packages = [{include = "event_planner_fastapi"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "0.101.0"
uvicorn = "^0.23.2"
pydantic = {version = "2.1.1", extras = ["email"]}
pydantic-settings = "2.0.2"
SQLAlchemy = "2.0.19"
passlib = {version = "1.7.4", extras = ["bcrypt"]}
aiomysql = "0.2.0"
cryptography = "41.0.3"
python-jose = {version = "3.3.0", extras = ["cryptography"]}
python-multipart = "0.0.6"
python-dotenv = "1.0.0"
greenlet = "2.0.2"


[tool.poetry.dev-dependencies]
pytest = "7.4.0"
httpx = "0.24.1"
pytest-asyncio = "0.21.1"
coverage = '7.3.0'


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"